+++ tools/diff-generator/src/lib/file-import.js
@@ -10,8 +10,12 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 import { access, readFile } from "fs/promises";
+import { existsSync } from "fs";
+import path from "path";
+import { glob } from "glob";
 
-const source = "https://raw.githubusercontent.com/adobe/spectrum-tokens/";
+const source = "https://raw.githubusercontent.com/";
+const defaultRepo = "adobe/spectrum-tokens/";
 
 /**
  * Returns file with given file name as a JSON object (took this from diff.js)
@@ -23,34 +27,108 @@ export default async function fileImport(
   givenTokenNames,
   givenVersion,
   givenLocation,
+  givenRepo,
+  githubAPIKey,
 ) {
   const version = givenVersion || "latest";
   const location = givenLocation || "main";
-  const tokenNames =
-    givenTokenNames || (await fetchTokens("manifest.json", version, location));
-  if (givenVersion === "test") {
-    await access(tokenNames);
-    return JSON.parse(await readFile(tokenNames, { encoding: "utf8" }));
-  }
   const result = {};
+  const tokenNames =
+    givenTokenNames ||
+    (await fetchTokens(
+      "manifest.json",
+      version,
+      location,
+      givenRepo,
+      githubAPIKey,
+    ));
   for (let i = 0; i < tokenNames.length; i++) {
-    const tokens = await fetchTokens(tokenNames[i], version, location);
+    const name = givenTokenNames ? "src/" + tokenNames[i] : tokenNames[i];
+    const tokens = await fetchTokens(
+      name,
+      version,
+      location,
+      givenRepo,
+      githubAPIKey,
+    );
     Object.assign(result, tokens);
   }
   return result;
 }
 
-async function fetchTokens(tokenName, version, location) {
+export async function loadLocalData(dirName, tokenNames) {
+  try {
+    const startDir = process.cwd();
+    const root = getRootPath(startDir, "pnpm-lock.yaml");
+    const fileNames = await glob(`${dirName}/*.json`, {
+      ignore: ["node_modules/**", "coverage/**"],
+      cwd: "../../",
+    }); // i.e. packages/tokens/src
+    return tokenNames
+      ? loadData(
+          root.substring(0, root.lastIndexOf("/")) + "/" + dirName + "/",
+          tokenNames,
+        )
+      : loadData(root.substring(0, root.lastIndexOf("/")) + "/", fileNames);
+  } catch (error) {
+    console.error(error);
+  }
+}
+
+async function loadData(startDir, tokenNames) {
+  let result = {};
+  for (let i = 0; i < tokenNames.length; i++) {
+    const tokenPath =
+      startDir + tokenNames[i].trim().replaceAll('"', "").replace(",", "");
+    await access(tokenPath);
+    const temp = JSON.parse(await readFile(tokenPath, { encoding: "utf8" }));
+    Object.assign(result, temp);
+  }
+  return result;
+}
+
+function getRootPath(startDir, targetDir) {
+  let curDir = startDir;
+  while (existsSync(curDir)) {
+    const curDirPath = path.join(curDir, targetDir);
+    if (existsSync(curDirPath)) {
+      return curDirPath;
+    }
+    const parentDir = path.dirname(curDir);
+    if (parentDir === curDir) {
+      return null;
+    }
+    curDir = parentDir;
+  }
+}
+
+async function fetchTokens(tokenName, version, location, repo, githubAPIKey) {
+  const repoURL = source + "/" + (repo && repo.length ? repo : defaultRepo);
   const link =
-    version !== "latest"
-      ? source + version.replace("@", "%40")
-      : source + location;
-  return (await fetch(`${link}/packages/tokens/${tokenName}`))
-    .json()
-    .then((tokens) => {
-      return tokens;
-    })
-    .catch((e) => {
-      console.log(e);
-    });
+    version !== "latest" ? repoURL + "/" + version : repoURL + "/" + location;
+
+  const url = `${link}/packages/tokens/${tokenName}`.replaceAll("//", "/");
+  const result = await fetch(
+    url,
+    githubAPIKey && githubAPIKey.length
+      ? {
+          headers: {
+            Authorization: "Bearer " + githubAPIKey, // api is rate limited without a personal access token
+          },
+        }
+      : {},
+  );
+
+  if (result && result.status === 200) {
+    return result
+      .json()
+      .then((tokens) => {
+        return tokens;
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  } else {
+    throw new Error(url + "\n\t" + result.status + ": " + result.statusText);
+  }
 }