+++ tools/diff-generator/test/fileImport.test.js
@@ -36,7 +36,7 @@ test("checking file import for two branches (both main)", async (t) => {
             reject(error);
           }
         })
-        .run("pnpm tdiff report -otb main -ntb main")
+        .run("./src/lib/cli.js report -otb main -ntb main")
         .end(resolve);
     } catch (error) {
       reject(error);
@@ -44,7 +44,7 @@ test("checking file import for two branches (both main)", async (t) => {
   });
 });
 
-test("checking cli output for two branches (shirlsli/diff-generator-cli and shirlsli/file-import-tests) for added color-component tokens", async (t) => {
+test("checking cli output for two branches (shirlsli/file-import-tests-1 and shirlsli/file-import-tests) for added color-component tokens", async (t) => {
   t.plan(1);
   return new Promise((resolve, reject) => {
     try {
@@ -62,7 +62,7 @@ test("checking cli output for two branches (shirlsli/diff-generator-cli and shir
           }
         })
         .run(
-          "pnpm tdiff report -otb shirlsli/diff-generator-cli -ntb shirlsli/file-import-tests -tn src/color-component.json",
+          "./src/lib/cli.js report -otb shirlsli/file-import-tests-1 -ntb shirlsli/file-import-tests -tn src/color-component.json",
         )
         .end(resolve);
     } catch (error) {
@@ -71,7 +71,7 @@ test("checking cli output for two branches (shirlsli/diff-generator-cli and shir
   });
 });
 
-test("checking cli output for two branches (shirlsli/diff-generator-cli and shirlsli/file-import-tests) for deleted layout tokens", async (t) => {
+test("checking cli output for two branches (shirlsli/file-import-tests-1 and shirlsli/file-import-tests) for deleted layout tokens", async (t) => {
   t.plan(1);
   return new Promise((resolve, reject) => {
     try {
@@ -89,7 +89,7 @@ test("checking cli output for two branches (shirlsli/diff-generator-cli and shir
           }
         })
         .run(
-          "pnpm tdiff report -otb shirlsli/diff-generator-cli -ntb shirlsli/file-import-tests -tn src/layout.json",
+          "./src/lib/cli.js report -otb shirlsli/file-import-tests-1 -ntb shirlsli/file-import-tests -tn src/layout.json",
         )
         .end(resolve);
     } catch (error) {
@@ -98,7 +98,7 @@ test("checking cli output for two branches (shirlsli/diff-generator-cli and shir
   });
 });
 
-test("checking cli output for two branches (shirlsli/diff-generator-cli and shirlsli/file-import-tests) for renamed, deprecated, added, deleted, and updated tokens", async (t) => {
+test("checking cli output for two branches (shirlsli/file-import-tests-1 and shirlsli/file-import-tests) for renamed, deprecated, added, deleted, and updated tokens", async (t) => {
   t.plan(1);
   return new Promise((resolve, reject) => {
     try {
@@ -116,7 +116,7 @@ test("checking cli output for two branches (shirlsli/diff-generator-cli and shir
           }
         })
         .run(
-          "pnpm tdiff report -otb shirlsli/diff-generator-cli -ntb shirlsli/file-import-tests",
+          "./src/lib/cli.js report -otb shirlsli/file-import-tests-1 -ntb shirlsli/file-import-tests",
         )
         .end(resolve);
     } catch (error) {
@@ -143,7 +143,7 @@ test("checking cli output between a release and a branch (@adobe/spectrum-tokens
           }
         })
         .run(
-          "pnpm tdiff report -otv @adobe/spectrum-tokens@12.26.0 -ntb shirlsli/file-import-tests",
+          "./src/lib/cli.js report -otv @adobe/spectrum-tokens@12.26.0 -ntb shirlsli/file-import-tests",
         )
         .end(resolve);
     } catch (error) {
@@ -170,7 +170,7 @@ test("checking cli output between a branch and a release (shirlsli/file-import-t
           }
         })
         .run(
-          "pnpm tdiff report -y -otb shirlsli/file-import-tests -ntv @adobe/spectrum-tokens@12.26.0",
+          "./src/lib/cli.js report -y -otb shirlsli/file-import-tests -ntv @adobe/spectrum-tokens@12.26.0",
         )
         .end(resolve);
     } catch (error) {